+++ src/rbtree.h
@@ -30,9 +30,10 @@
 #define	_LINUX_RBTREE_H
 
 #include "util.h"
+#include <stdint.h>
 
 struct rb_node {
-	unsigned long  __rb_parent_color;
+	uintptr_t  __rb_parent_color;
 	struct rb_node *rb_right;
 	struct rb_node *rb_left;
 };
@@ -51,7 +52,7 @@ extern void rb_erase(struct rb_node *, struct rb_root *);
 static inline void rb_link_node(struct rb_node * node, struct rb_node * parent,
 				struct rb_node ** rb_link)
 {
-	node->__rb_parent_color = (unsigned long)parent;
+	node->__rb_parent_color = (uintptr_t)parent;
 	node->rb_left = node->rb_right = NULL;
 
 	*rb_link = node;
